feat: about section buttons
