+++ client/src/app/sections/utils/UASCLogo.tsx
@@ -1,0 +1,36 @@
+import { SVGProps } from "react"
+
+const UASCLogo = (props: SVGProps<SVGSVGElement>) => {
+  return (
+    <svg
+      width="116"
+      height="99"
+      viewBox="0 0 116 99"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <g clipPath="url(#clip0_5162_7448)">
+        <path
+          d="M0.189883 99C0.189883 97.28 0.189883 95.56 0.129883 93.85C0.129883 93.07 0.439883 92.74 1.19988 92.77H2.38988C14.8799 92.77 27.3899 92.77 39.8599 92.69C41.3693 92.6522 42.8706 92.4578 44.3399 92.11C44.8997 92.0095 45.3987 91.6961 45.7325 91.2356C46.0662 90.7751 46.2086 90.2032 46.1299 89.64C46.0799 86.64 46.1299 83.64 46.1299 80.64C46.1299 79.18 45.6899 78.4 44.2799 78.09C42.5785 77.6834 40.8387 77.4588 39.0899 77.42C30.5099 77.42 21.9299 77.42 13.3499 77.42C10.304 77.5205 7.27675 76.9075 4.50988 75.63C2.05988 74.42 0.429883 72.54 0.289883 69.78C0.139881 66.8319 0.139881 63.8781 0.289883 60.93C0.33138 59.7742 0.702757 58.6544 1.36016 57.7029C2.01756 56.7513 2.93352 56.0078 3.99988 55.56C6.04939 54.7013 8.18488 54.0644 10.3699 53.66C11.987 53.4093 13.625 53.3189 15.2599 53.39C25.6332 53.35 36.0065 53.35 46.3799 53.39C48.0205 53.4754 49.6514 53.0907 51.0809 52.2811C52.5105 51.4716 53.6792 50.2709 54.4499 48.82C63.1166 34.08 71.8299 19.36 80.5899 4.66C81.5199 3.07 82.4699 1.54 83.3999 0H83.8799C84.5699 1.21 85.2599 2.42 85.9599 3.62C94.4799 18.0933 103 32.5533 111.52 47C112.76 49.1 114.06 51.15 115.33 53.23V54C113.01 54 110.7 54 108.33 54C107.957 54.0298 107.585 53.9402 107.267 53.7442C106.949 53.5483 106.701 53.2561 106.56 52.91C104.14 48.55 101.67 44.2 99.2099 39.91C99.0699 39.66 98.8699 39.44 98.5899 39.05C98.3999 40.18 98.2899 41.05 98.0599 41.98C97.839 42.6599 97.582 43.3275 97.2899 43.98C96.6837 43.5605 96.0995 43.1099 95.5399 42.63C94.5399 41.63 93.5399 40.63 92.3799 39.45C92.2599 40.45 92.1299 41.18 92.1099 41.87C91.9999 45.25 91.9999 48.64 91.7599 52C91.5781 52.7701 91.267 53.5039 90.8399 54.17C90.2825 53.6426 89.8012 53.0401 89.4099 52.38C87.7599 48.61 86.2099 44.8 84.5899 41.01C83.5899 38.59 82.4899 36.18 81.3599 33.8C80.8199 32.68 80.2899 32.68 79.7199 33.8C78.9199 35.36 78.2099 36.98 77.5499 38.6C76.3499 41.52 75.2299 44.48 74.0299 47.39C73.6783 48.0846 73.2773 48.753 72.8299 49.39L72.3399 49.29C70.8599 45.29 70.5299 41.12 69.6099 36.74C69.2299 37.35 68.9699 37.74 68.7399 38.13C65.9999 43.07 63.2399 48 60.3499 52.85C57.7299 57.22 54.0199 59.66 48.6599 59.61C36.8499 59.48 25.0399 59.61 13.2199 59.61C11.9505 59.6797 10.6961 59.9185 9.48988 60.32C9.00546 60.4413 8.58394 60.7395 8.30821 61.1558C8.03249 61.5722 7.9225 62.0766 7.99988 62.57C8.05988 64.36 7.99988 66.16 7.99988 67.96C7.99988 69.76 8.40988 70.39 10.1399 70.77C11.6934 71.1146 13.2787 71.2956 14.8699 71.31C23.4499 71.31 32.0299 71.31 40.6099 71.31C43.5748 71.2107 46.525 71.7673 49.2499 72.94C50.6585 73.4892 51.8602 74.465 52.6868 75.7309C53.5134 76.9968 53.9235 78.4895 53.8599 80C53.8599 83.56 53.8599 87.12 53.8599 90.68C53.9022 91.804 53.6471 92.9193 53.1204 93.9131C52.5936 94.9069 51.8138 95.7441 50.8599 96.34C49.7435 97.0657 48.5274 97.6249 47.2499 98C45.6599 98.46 43.9799 98.65 42.3399 99H0.189883ZM73.8399 39.35C74.1399 39.07 74.3699 38.94 74.4499 38.75C74.9599 37.67 75.4499 36.58 75.9099 35.47C77.0999 32.55 78.2399 29.6 79.4699 26.7C79.6599 26.24 80.1899 25.92 80.5699 25.54C80.9808 25.8852 81.338 26.2896 81.6299 26.74C82.3699 28.41 83.0399 30.13 83.7099 31.84C85.1399 35.48 86.5299 39.14 87.9799 42.78C88.2797 43.3392 88.642 43.8626 89.0599 44.34L89.4399 44.21C89.5524 43.6862 89.6259 43.1547 89.6599 42.62C89.6599 39.87 89.6599 37.11 89.6599 34.35C89.6652 33.8987 89.6953 33.448 89.7499 33L90.1199 32.79L95.2099 37.87C95.6099 36.05 96.3199 34.64 95.4799 33.14C91.7332 26.4733 87.9666 19.8067 84.1799 13.14C84.0035 12.871 83.8098 12.6138 83.5999 12.37C83.3699 12.68 83.2099 12.86 83.0899 13.06C79.5899 19.1733 76.0932 25.3033 72.5999 31.45C72.3667 31.8199 72.2548 32.2534 72.2799 32.69C72.5499 34.47 72.8799 36.23 73.2799 37.99C73.4183 38.4617 73.606 38.9176 73.8399 39.35Z"
+          fill="currentColor"
+        />
+        <path
+          d="M115.34 64.52C102.26 64.52 89.1799 64.52 76.0999 64.62C74.9076 64.647 73.7255 64.8457 72.5899 65.21C71.955 65.3387 71.3938 65.706 71.0217 66.2362C70.6496 66.7664 70.495 67.4192 70.5899 68.06C70.5899 75.1266 70.5899 82.2 70.5899 89.28C70.5899 90.82 71.1399 91.76 72.5899 92.12C74.2847 92.5474 76.0224 92.7822 77.7699 92.82C89.7699 92.82 101.69 92.82 113.66 92.82C114.21 92.82 114.77 92.82 115.33 92.82V98.82H73.4099C71.8599 98.46 70.2799 98.22 68.7699 97.72C64.5899 96.35 62.6399 93.59 62.6299 89.19C62.6299 82.15 62.6299 75.11 62.6299 68.07C62.6299 63.62 64.6299 60.64 68.9699 59.45C71.5437 58.7354 74.1989 58.3556 76.8699 58.32C89.6799 58.18 102.49 58.23 115.3 58.21L115.34 64.52Z"
+          fill="currentColor"
+        />
+        <path
+          d="M0 21.58C0 16.67 0 11.75 0 6.83004C0 5.67004 0.35 5.27004 1.5 5.33004C3.09553 5.41508 4.69447 5.41508 6.29 5.33004C7.44 5.27004 7.79 5.67004 7.79 6.83004C7.79 15.94 7.79 25.06 7.79 34.18C7.79 37.95 9.32 39.58 13.14 39.66C16.96 39.74 21.04 39.66 24.99 39.66C28.94 39.66 32.99 39.66 36.99 39.66C37.5856 39.65 38.1786 39.5796 38.76 39.45C41.65 38.82 42.68 37.56 42.68 34.61C42.68 25.57 42.68 16.5334 42.68 7.50004C42.68 7.14004 42.68 6.78004 42.68 6.42004C42.62 5.58004 43.03 5.29004 43.83 5.31004C45.66 5.31004 47.5 5.31004 49.33 5.31004C50.18 5.31004 50.44 5.66004 50.44 6.46004C50.44 16.38 50.51 26.29 50.44 36.21C50.4814 38.1985 49.8339 40.14 48.6073 41.7056C47.3807 43.2711 45.6505 44.3644 43.71 44.8C41.4109 45.4529 39.0392 45.8157 36.65 45.88C29 46 21.36 46 13.74 45.9C10.6936 46.0196 7.66451 45.3877 4.92 44.06C3.4481 43.3528 2.20985 42.2383 1.35202 40.8487C0.494186 39.4592 0.0527376 37.8528 0.08 36.22C0 31.34 0.05 26.46 0 21.58Z"
+          fill="currentColor"
+        />
+      </g>
+      <defs>
+        <clipPath id="clip0_5162_7448">
+          <rect width="115.39" height="98.99" fill="white" />
+        </clipPath>
+      </defs>
+    </svg>
+  )
+}
+
+export default UASCLogo
