+++ client/src/components/generic/FigmaButtons/FigmaButton.tsx
@@ -1,5 +1,5 @@
-import RightArrowIcon from "@/assets/icons/rightarrow.svg"
 import LeftArrowIcon from "@/assets/icons/leftarrow.svg"
+import RightArrowIcon from "@/assets/icons/rightarrow.svg"
 
 type buttonVariants =
   | "default"
